Move StringComparers into Utils namespace
