+++ src/Aspire.Dashboard/Components/Pages/ConsoleLogs.razor.cs
@@ -3,6 +3,7 @@
 
 using Aspire.Dashboard.ConsoleLogs;
 using Aspire.Dashboard.Model;
+using Aspire.Dashboard.Utils;
 using Microsoft.AspNetCore.Components;
 using Microsoft.FluentUI.AspNetCore.Components;
 using Microsoft.JSInterop;
