+++ src/Aspire.Dashboard/Model/ContainerViewModel.cs
@@ -2,7 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 
 using System.Collections.Immutable;
-using Aspire.Dashboard.Extensions;
+using Aspire.Dashboard.Utils;
 
 namespace Aspire.Dashboard.Model;
