+++ src/Aspire.Dashboard/Utils/StringComparers.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Aspire.Dashboard.Extensions;
+namespace Aspire.Dashboard.Utils;
 
 internal static class StringComparers
 {
